+++ spec/fixtures/WorkspaceSchemes/GroupLocation/GroupTypeGroupLocation/project_in_group_type_group/project_in_group_type_group/ViewController.swift
@@ -1,0 +1,20 @@
+//
+//  ViewController.swift
+//  project_in_group_type_group
+//
+//  Created by kezi on 2019/02/06.
+//  Copyright © 2019 CocoaPods. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view, typically from a nib.
+    }
+
+
+}
+