+++ spec/workspace_spec.rb
@@ -150,7 +150,10 @@ describe Xcodeproj::Workspace do
     end
 
     it 'contains only test data schemes' do
-      @workspace.schemes.keys.sort.should == %w(WorkspaceSchemesApp WorkspaceSchemesFramework WorkspaceSchemesScheme project_in_subgroup scheme_in_subgroup_with_location)
+      @workspace.schemes.keys.sort.should == %w(
+        WorkspaceSchemesApp WorkspaceSchemesFramework WorkspaceSchemesScheme
+        project_in_group_type_group project_in_subgroup scheme_in_subgroup_with_location
+      )
     end
 
     it 'schemes hash contain path to a valid project/workspace' do