+++ CHANGELOG.md
@@ -16,6 +16,11 @@
   [joshdholtz](https://github.com/joshdholtz)
   [#656](https://github.com/CocoaPods/CocoaPods/pull/656)  
 
+* Ensure a `GroupReference`'s path includes its parent `GroupReference`'s path. 
+  Both `FileReference`s and `GroupReference`s only prepend the parent path if
+  the child has a type of `group`.
+  [Kesi Maduka](https://github.com/k3zi)
+  [#657](https://github.com/CocoaPods/Xcodeproj/issues/657)
 
 ## 1.8.0 (2019-01-25)